requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-groq>=0.1.0
langgraph>=0.2.0
openai>=1.50.0
chromadb>=0.5.0
streamlit>=1.39.0
plotly>=5.24.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
tiktoken>=0.8.0
pydantic>=2.9.0
numpy>=1.26.0
